Move `incident.spec` to `src/test`
